LayoutTests/storage/indexeddb/basics-expected.txt

Test the basics of IndexedDB.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'indexedDB' in window is true
PASS 'open' in indexedDB is true
indexedDB.open('name', 'description', true /* allow modification */)
PASS 'onsuccess' in result is true
PASS 'onerror' in result is true
PASS 'abort' in result is true
PASS 'readyState' in result is true

An event should fire shortly...

Success function called

PASS 'source' in event is true
PASS 'open' in event.source is true

PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false

PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'abort' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

PASS successfullyParsed is true

TEST COMPLETE
